PF/PB/requirements.txt
django == 4.0.1
pillow == 8.2.0
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.2.2
python-dateutil==2.8.2
django-google-maps == 0.13.0
geopy == 2.3.0
